服务网关调用评论接口

+++ community-service/src/main/java/com/hzu/community/controller/UserController.java
@@ -22,7 +22,7 @@ public class UserController {
     private UserService userService;
 //    查询
     @RequestMapping("getById")
-    public RestResponse<User> getUserById(Long id){
+    public RestResponse<User> getUserById(Integer id){
        User user= userService.getUserById(id);
             return RestResponse.success(user);
     }
